requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
huggingface_hub==0.26.5
pydantic==2.10.4
duckduckgo-search==6.3.7
wikipedia==1.4.0
pypdf==5.1.0
sentence-transformers==3.3.1
chromadb==0.5.23
langchain-community==0.3.9
langchain-text-splitters==0.3.2
python-multipart==0.0.17
python-dotenv>=1.0.0
tf-keras>=2.20.0
